chore: Update README.md with license information for code and TTF files
The README.md file has been updated to include license information for the code and TTF files used in the project. The code and *.pth files are under the MIT License, while the TTF files in the `fonts/` directory are copied from [kiwi0fruit/open-fonts](https://github.com/kiwi0fruit/open-fonts) and are subject to their own license.